Move SendCopyToEmail component to shared internal folder
