+++ packages/lib/src/components/internal/SendCopyToEmail/SendCopyToEmail.tsx
@@ -1,9 +1,9 @@
 import { h } from 'preact';
 import cx from 'classnames';
 import { useState } from 'preact/hooks';
-import { renderFormField } from '../../../internal/FormFields';
-import useCoreContext from '../../../../core/Context/useCoreContext';
-import Field from '../../../internal/FormFields/Field';
+import { renderFormField } from '../FormFields';
+import useCoreContext from '../../../core/Context/useCoreContext';
+import Field from '../FormFields/Field';
 
 export default function SendCopyToEmail(props) {
     const { errors, value, onInput, onChange } = props;
@@ -29,7 +29,7 @@ export default function SendCopyToEmail(props) {
             {sendCopyToEmail && (
                 <Field label={i18n.get('shopperEmail')} classNameModifiers={['shopperEmail']} errorMessage={errors}>
                     {renderFormField('emailAddress', {
-                        name: 'boleto.shopperEmail',
+                        name: 'shopperEmail',
                         autoCorrect: 'off',
                         spellCheck: false,
                         value,
