+++ packages/lib/src/components/internal/SendCopyToEmail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SendCopyToEmail';
